+++ src/app/admin/lodge/create/page.tsx
@@ -1,11 +1,12 @@
 "use client";
 
 import React from "react";
-import { createLodge, RoomType } from "@/lib/admin/lodge/lodgeThunk";
+import { createLodge } from "@/lib/admin/lodge/lodgeThunk";
 import { useAppDispatch } from "@/lib/store/hooks";
 import { useRouter } from "next/navigation";
 import LodgeForm from "@/components/ui/LodgeForm";
 import { ArrowLeft } from "lucide-react";
+import { RoomType } from "@/types/lodge";
 
 const CreateLodgePage = () => {
   const dispatch = useAppDispatch();
